refactor: change the naming convension for files and directories
