--- app/_Home/index.ts
@@ -1,1 +1,0 @@
-export * from "./Hero"
