+++ components/Neon/index.ts
@@ -1,2 +1,2 @@
-export * from "./halfRoundMesh"
-export * from "./RoundMesh"
+export * from "./half_round_mesh"
+export * from "./round_mesh"
