+++ components/aceternity/index.ts
@@ -1,1 +1,1 @@
-export * from "./Spotlight"
+export * from "./spotlight"
