+++ components/onBoarding/boarding_floating_bar.tsx
