+++ components/site_header.tsx
@@ -3,10 +3,10 @@ import Link from "next/link"
 import { siteConfig } from "@/config/site"
 import { buttonVariants } from "@/components/ui/button"
 import { Icons } from "@/components/icons"
-import { MainNav } from "@/components/main-nav"
-import { ThemeToggle } from "@/components/theme-toggle"
+import { MainNav } from "@/components/main_nav"
+import { ThemeToggle } from "@/components/theme_toggle"
 
-import HamMenu from "./hamMenu"
+import HamMenu from "./ham_menu"
 
 export function SiteHeader() {
   return (
