+++ hooks/index.ts
@@ -1,1 +1,1 @@
-export * from "./useDebounce"
+export * from "./use_debounce"
