+++ types/model/index.ts
@@ -6,4 +6,4 @@
  * OpenAPI spec version: 0.0.1
  */
 
-export * from "./modelsCat"
+export * from "./models_cat"
